Added .cpp files to fix multiple definitions
