+++ main.ino
@@ -51,26 +51,3 @@ void countLeft() {
 void countRight() {
   encoderValueRight++;
 }
-
-void printEncoderValues() {
-  Serial1.print("Encoder Value: ");
-  Serial1.println(encoderValueLeft);
-}
-
-void checkIfTooClose() {
-  if (isTooClose()) {
-    while (true) {
-      Serial1.println("Stopped: Too close");
-      blink(1);
-    }
-  }
- }
-
-bool isTooClose() {
-  sensors.readSensors();
-  if (sensors.getFrontPTReading() > 964) {  // Prevent motor driver from burning out
-    motors.brake();
-    return true;
-  }
-  return false;
-}
